[BUGFIX] Improve type and object version handling when deserializing
Addresses issues with type and differing versions of the `SiteLanguage` class when deserializing `RecordInstanceIdentifier` objects in `ReNamespaceDeferredOperationsUpdateWizard`.

How locale information is stored in `SiteLanguage` has changed between different TYPO3 versions. This fix only allows RecordInstanceIdentifier and RecordRepresentation objects to be deserialized. All other objects are deserialized as __PHP_Incomplete_Class and the object properties can be accessed without reliance on the current TYPO3 version being compatible with the serialized `SiteLanguage` and `Locale` objects.

For example, `SiteLanguage::$locale` was a string in TYPO3 <=v12 and a `Locale` object in later versions, which led to a "Cannot assign string to property" exception when deserializing older objects.

This bug fix also allows the upgrade wizard to handle deferred operations serialized directly as `RecordRepresentation` objects. Newer serializations use an array with both `RecordRepresentation` and metadata.

Resolves: #129